Renamed variable to clean up naming conventions.
