+++ mgl-webapp/src/main/java/com/organization/mvcproject/model/Review.java
@@ -1,0 +1,32 @@
+package com.organization.mvcproject.model;
+
+public class Review {
+
+	private String body;
+	private String author;
+	private Integer rating;
+
+	public String getAuthor() {
+		return author;
+	}
+
+	public void setAuthor(String author) {
+		this.author = author;
+	}
+
+	public Integer getRating() {
+		return rating;
+	}
+
+	public void setRating(Integer rating) {
+		this.rating = rating;
+	}
+
+	public String getBody() {
+		return body;
+	}
+
+	public void setBody(String reviewBody) {
+		this.body = reviewBody;
+	}
+}
